Implementing router v1.0 (#36)
Accept requirements as args (e.g: /users/{id:\d+}).

Request and Response are sent automatically if the controller use them.

PHP routes configuration file

Implement UrlGeneratorInterface.

Implement all tests in RouterTest.php, it must be run after all modifications.

The index file is the entry point. It generates the router then loads the routes inside it. Next, the router->call() is called with the Request object. If a route matches, its controller is called and it returns a Response. After all, the response is sent from the index.

When loading a route, we can pass parameters and expected regex, e.g: '/users/{id:\d+}' will look only for digits as the id parameter. If the parameter is incorrect, it returns a RouteNotFoundException.