Implements dependency injection container and tests
We can retrieve our services automatically with the $container->get() method. If the service needs dependencies, they are also autowired.

When requesting a service from the container, it can be the same instance for all call, or a new one by specifying the $container->getDefinition(MyService::class)->setShared(false) before asking for it

Create Kernel class used in the web application entrypoint (public/index.php)