Dependency injection container (#37)
* Add DIC, update the kernel, route and router.

The order of the routes definitions is no longer important. The response is not sent by the router anymore, it's now the kernel that loads and call the router, get and call a callable provided by the router, then send a response.

Autowiring was added to the controllers methods through the kernel->run() method, using the DIC.

Refactor tests.