Work in progress on models
executeInsertions method works, and we can insert some entities with $em->add($entity) and $em->flush()

The DataMapper read the attached attributes of an entity and create a Metadata instance through its getMetadata() method.

The UnitOfWork executes the queued insertions with an SQL transaction and rollback changes if an error occurs.