Refactor constraints, validators and Request handling. Create a FileUploader service. Refactor object handling from the hydrator.
A constraint attribute now get its own constraint validator which is called automatically by the Validator. It's useful to initialize some required objects from the constructor (like the MySQLAdapter in the Unique constraint validation).

A FileUploader is now used to upload or replace a file.

The main Request is now initialized when configuring the DIC: In this way, we can retrieve the main request everywhere we need it (like when generating the csrf_token from the Csrf Twig Extension).

Add many tests.