Applying changes
Replace Password Strength constraint validator because it was not what we expected at first.

Reset Password feature: Add a renderView method inside the abstract controller to load a view without returning something. In this way we can load the new mails/* views.

There was an unexpected behavior from the Templating and the Abstract Controller method about the flashes messages. Now, we specify if we are in the main request in case we want to call something else (e.g renderView()) without clearing the flashBag from the session.

Add error 500 handling and Unauthorized exception redirecting the user to the login page.